style: rename directories
