+++ src/test/java/bowling_step3/FrameGeneralTest.java
@@ -1,4 +1,4 @@
-package bowling_min;
+package bowling_step3;
 
 import bowling_step3.domain.*;
 import bowling_step3.view.Output;
