+++ src/test/java/bowling_step3/ScoresTest.java
@@ -1,4 +1,4 @@
-package bowling_min;
+package bowling_step3;
 
 import bowling_step3.domain.Scores;
 import org.junit.jupiter.api.Test;
